+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [Unreleased]
+
+### Fix
+
+- Avoid false T002 reports for template tags inside HTML attributes.
+
 ## [1.40.7] - 2026-07-15
 
 ### Fix